backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
chromadb==0.4.22
langchain==0.0.354
langchain-chroma==0.0.4
langchain-community==0.0.10
sentence-transformers==2.2.2
PyPDF2==3.0.1
httpx==0.25.1